create admin users component
