+++ frontend/src/components/pages/AdminPage.tsx
@@ -15,12 +15,10 @@ import {
 import { useQuery } from "@apollo/client";
 
 import { AdminUser } from "../../types/UserTypes";
-import AdminUserTable from "../user-management/AdminUserTable";
 import AddAdminModal from "../user-management/AddAdminModal";
 import { AlertIcon } from "../../assets/icons";
 import { GET_USERS_BY_ROLE } from "../../APIClients/queries/UserQueries";
-import SortTablePopover from "../common/SortTablePopover";
-import SearchBar from "../common/SearchBar";
+import AdminUsers from "../user-management/AdminUsers";
 import AdminDashboard from "../common/admin/AdminDashboard";
 
 const ErrorState = (): React.ReactElement => (
@@ -41,56 +39,18 @@ const getAdminUser = (user: AdminUser) => {
   };
 };
 
-type AdminUserProperty = "firstName" | "email";
-type SortOrder = "Ascending" | "Descending";
-
 const AdminPage = (): React.ReactElement => {
   const unselectedColor = useColorModeValue("#727278", "#727278");
-  const [search, setSearch] = React.useState("");
-  const [sortProperty, setSortProperty] = React.useState<AdminUserProperty>(
-    "firstName",
-  );
-  const [sortOrder, setSortOrder] = React.useState<SortOrder>("Ascending");
 
-  const OrderingSets = {
-    sortProperty,
-    sortOrder,
-    setSortProperty,
-    setSortOrder,
-  };
-
-  const { loading, error, data } = useQuery(GET_USERS_BY_ROLE, {
+  const { loading, error, data: adminData } = useQuery(GET_USERS_BY_ROLE, {
     fetchPolicy: "cache-and-network",
     variables: { role: "Admin" },
   });
 
-  const filteredAdmins = React.useMemo(() => {
-    let filteredUsers = data?.usersByRole;
-    if (search) {
-      filteredUsers = filteredUsers.filter(
-        (user: AdminUser) =>
-          `${user.firstName} ${user.lastName}`
-            .toLowerCase()
-            .includes(search.toLowerCase()) ||
-          user.email.toLowerCase().includes(search.toLowerCase()),
-      );
-    }
-    return filteredUsers?.map(getAdminUser);
-  }, [search, data]);
-
   const admins = React.useMemo(() => {
-    let sortedUsers: AdminUser[] = filteredAdmins as AdminUser[];
-    if (sortOrder === "Descending") {
-      sortedUsers = sortedUsers?.sort((a, b) =>
-        a[sortProperty].toLowerCase() < b[sortProperty].toLowerCase() ? 1 : -1,
-      );
-    } else if (sortOrder === "Ascending") {
-      sortedUsers = sortedUsers?.sort((a, b) =>
-        a[sortProperty].toLowerCase() > b[sortProperty].toLowerCase() ? 1 : -1,
-      );
-    }
-    return sortedUsers;
-  }, [filteredAdmins, sortProperty, sortOrder]);
+    const filteredUsers = adminData?.usersByRole;
+    return filteredUsers?.map(getAdminUser);
+  }, [adminData]);
 
   const page = (
     <>
@@ -117,7 +77,7 @@ const AdminPage = (): React.ReactElement => {
           <ErrorState />
         </Center>
       )}
-      {data && !error && !loading && (
+      {adminData && !error && !loading && (
         <Box flex="1">
           <Tabs marginTop={3}>
             <TabList>
@@ -126,21 +86,10 @@ const AdminPage = (): React.ReactElement => {
             </TabList>
             <TabPanels>
               <TabPanel>
-                <VStack pt={4} spacing={6}>
-                  <HStack width="100%">
-                    <SearchBar onSearch={setSearch} />
-                    <SortTablePopover OrderingSets={OrderingSets} />
-                  </HStack>
-                  {search && (
-                    <Text fontSize="16px" color="grey.300" width="100%">
-                      Showing {admins.length} results for &quot;{search}&quot;
-                    </Text>
-                  )}
-                  <AdminUserTable adminUsers={admins} />
-                </VStack>
+                <AdminUsers admins={admins} />
               </TabPanel>
               <TabPanel>
-                <AdminUserTable adminUsers={admins} />
+                <p>hi</p>
               </TabPanel>
             </TabPanels>
           </Tabs>
